collections: use stream-json to handle streaming
