+++ packages/collections/package.json
@@ -29,7 +29,8 @@
   ],
   "dependencies": {
     "@openfn/language-common": "workspace:*",
-    "jsonparse": "^1.3.1"
+    "stream-chain": "^3.3.2",
+    "stream-json": "^1.8.0"
   },
   "devDependencies": {
     "assertion-error": "2.0.0",
