+++ packages/collections/test/mock/server.test.js
@@ -1,120 +1,15 @@
 import { expect } from 'chai';
-import { Client, MockAgent } from 'undici';
-import Parser from 'jsonparse';
+import { createServer } from '../mock/server';
+import { streamResponse } from '../../src/Utils';
 
-import { createServer, API } from '../mock/server';
-
-const parser = new Parser();
-
-describe('API', () => {
-  let api;
-
-  before(() => {
-    api = new API();
-  });
-
-  afterEach(() => {
-    api.reset();
-  });
-
-  it('should create an API', () => {
-    expect(api).to.not.be.undefined;
-
-    expect(api.collections).to.eql({});
-  });
-
-  it('should create a collection', () => {
-    api.createCollection('a');
-
-    expect(api.collections).to.eql({ a: {} });
-  });
-
-  it('should reset', () => {
-    api.createCollection('x');
-
-    expect(api.collections).to.eql({ x: {} });
-
-    api.reset();
-
-    expect(api.collections).to.eql({});
-  });
-
-  it('should create two APIs with two different collections', () => {
-    const a = new API();
-    a.createCollection('a');
-
-    const b = new API();
-    b.createCollection('b');
-
-    expect(a.collections.b).to.be.undefined;
-    expect(b.collections.a).to.be.undefined;
-  });
-
-  it('should add to a collection', () => {
-    api.createCollection('a');
-
-    api.upsert('a', 'x', {});
-
-    expect(api.collections.a).to.eql({ x: {} });
-  });
-
-  it("should throw if adding to add to a collection that doesn't exist", () => {
-    try {
-      api.upsert('c', 'x', {});
-    } catch (e) {
-      expect(e.message).to.eql('COLLECTION_NOT_FOUND');
-    }
-  });
-
-  it('should replace into a collection', () => {
-    api.createCollection('a');
-
-    api.upsert('a', 'x', []);
-
-    expect(api.collections.a).to.eql({ x: [] });
-  });
-
-  it('should fetch from a collection', () => {
-    api.createCollection('a');
-
-    api.upsert('a', 'x', { id: 1 });
-
-    const result = api.fetch('a', 'x', {});
-
-    expect(result).to.eql([{ id: 1 }]);
-  });
-
-  it('should fetch from a collection with wildcard', () => {
-    api.createCollection('a');
-
-    api.upsert('a', 'x', { id: 1 });
-    api.upsert('a', 'xx', { id: 2 });
-    api.upsert('a', 'axb', { id: 3 });
-    api.upsert('a', 'yy', { id: 4 });
-
-    const { results } = api.fetch('a', 'x*', {});
-
-    expect(results).to.eql([{ id: 1 }, { id: 2 }, { id: 3 }]);
-  });
-
-  it('should remove from a collection', () => {
-    api.createCollection('a');
-
-    api.upsert('a', 'x', { id: 1 });
-
-    api.remove('a', 'x');
-
-    expect(api.collections.a).to.eql({});
-  });
-});
-
-describe.only('GET', () => {
+// TODO: support query options
+// TODO support timestamps
+describe('GET', () => {
   let request;
-  let stream;
   let api;
 
   beforeEach(() => {
-    ({ api, request, stream } = createServer());
+    ({ api, request } = createServer());
   });
 
   it('should return 200 for a valid collection', async () => {
@@ -129,28 +24,69 @@ describe.only('GET', () => {
     expect(response.statusCode).to.equal(404);
   });
 
-  // I think this is really a clientside problem?
-  it.only('should consume results as a stream', async done => {
+  it('should consume all results as a stream', async () => {
     api.createCollection('my-collection');
 
-    api.upsert('my-collection', 'x', { id: 'x ' });
-    api.upsert('my-collection', 'y', { id: 'y ' });
-    api.upsert('my-collection', 'z', { id: 'z ' });
+    api.upsert('my-collection', 'x', { id: 'x' });
+    api.upsert('my-collection', 'y', { id: 'y' });
+    api.upsert('my-collection', 'z', { id: 'z' });
 
     const response = await request('/collections/my-collection');
+    const results = [];
 
-    // hmm. The way this returns data is a little wierd
-    parser.onValue = function (value) {
-      console.log(this, value);
-    };
+    await streamResponse(response, item => results.push(item));
 
-    for await (const chunk of response.body) {
-      parser.write(chunk);
-    }
-    done();
+    expect(results).to.eql([
+      { key: 'x', value: { id: 'x' } },
+      { key: 'y', value: { id: 'y' } },
+      { key: 'z', value: { id: 'z' } },
+    ]);
+  });
+
+  it('should consume a single result as a stream', async () => {
+    api.createCollection('my-collection');
+
+    api.upsert('my-collection', 'x', { id: 'x' });
+    api.upsert('my-collection', 'y', { id: 'y' });
+    api.upsert('my-collection', 'z', { id: 'z' });
+
+    const response = await request('/collections/my-collection/y');
+    const results = [];
+
+    await streamResponse(response, item => results.push(item));
+
+    expect(results).to.eql([{ key: 'y', value: { id: 'y' } }]);
+  });
+
+  it('should consume results as a stream with a wildcard', async () => {
+    api.createCollection('my-collection');
+
+    api.upsert('my-collection', 'ax', { id: 'x' });
+    api.upsert('my-collection', 'ay', { id: 'y' });
+    api.upsert('my-collection', 'az', { id: 'z' });
+
+    const response = await request('/collections/my-collection/*z');
+    const results = [];
+
+    await streamResponse(response, item => results.push(item));
+
+    expect(results).to.eql([{ key: 'az', value: { id: 'z' } }]);
   });
 });
 
-describe('POST', () => {});
+// describe.only('POST', () => {
+//   it('should return 200 for a valid collection', async () => {
+//     api.createCollection('my-collection');
+
+//     const response = await request('/collections/my-collection');
+//     expect(response.statusCode).to.equal(200);
+//   });
+
+//   it("should return 404 for a collection that doesn't exist", async () => {
+//     const response = await request('/collections/my-collection');
+//     expect(response.statusCode).to.equal(404);
+//   });
+
+// });
 
 describe('DELETE', () => {});
